backend/requirements.txt
# Core Django
Django==6.0.4
asgiref==3.11.1
sqlparse==0.5.5

# Database & Environment
psycopg2-binary
python-decouple

# API & Security
djangorestframework
django-cors-headers
djangorestframework-simplejwt

# Scraping & Utilities
requests
beautifulsoup4
